requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-dotenv
cryptography
pydantic[email]
aiohttp>=3.9.5
